requirements.txt
numpy<2.0.0
pandas>=2.0.0
tqdm
pdfplumber
sentence-transformers==2.7.0
faiss-cpu==1.8.0
langchain>=0.2.0
langgraph>=0.1.0
langchain-community>=0.2.0
python-dotenv
streamlit
langchain-ollama
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
pydantic>=2.0.0
transformers>=4.40.0
groq

